// components/DetailsPage/enums.ts
export enum DetailsBlockTitle {
  Genres = "Genres",
  Casts = "Casts",
  SpokenLanguages = "Spoken Languages",
  ProductionCountries = "Production Countries",
  ProductionCompanies = "Production Companies",
  ReleaseDate = "Release Date",
  FirstAirDate = "First Air Date",
  LastAirDate = "Last Air Date",
  IMDB = "IMDB",
  Revenue = "Revenue",
  NumberOfSeasons = "Number of seasons",
  NumberOfEpisodes = "Number of episodes",
  Runtime = "Runtime(min)",
  Status = "Status",
}

export enum DetailsPageActionButtons {
  HomePage = "Home Page",
  GoBack = "Go Back",
  Play = "Play Trailer",
}